8286850: [macos] Add support for signing user provided app image

+++ test/jdk/tools/jpackage/helpers/jdk/jpackage/test/JPackageCommand.java
@@ -71,6 +71,7 @@ public final class JPackageCommand extends CommandArguments<JPackageCommand> {
         suppressOutput = cmd.suppressOutput;
         ignoreDefaultRuntime = cmd.ignoreDefaultRuntime;
         ignoreDefaultVerbose = cmd.ignoreDefaultVerbose;
+        noCleanupBeforeExec = cmd.noCleanupBeforeExec;
         immutable = cmd.immutable;
         prerequisiteActions = new Actions(cmd.prerequisiteActions);
         verifyActions = new Actions(cmd.verifyActions);
@@ -663,6 +664,12 @@ public final class JPackageCommand extends CommandArguments<JPackageCommand> {
         return this;
     }
 
+    public JPackageCommand noCleanupBeforeExec(boolean v) {
+        verifyMutable();
+        noCleanupBeforeExec = v;
+        return this;
+    }
+
     public boolean isWithToolProvider() {
         return Optional.ofNullable(withToolProvider).orElse(
                 defaultWithToolProvider);
@@ -702,13 +709,15 @@ public final class JPackageCommand extends CommandArguments<JPackageCommand> {
     public Executor.Result execute(int expectedExitCode) {
         executePrerequisiteActions();
 
-        if (isImagePackageType()) {
-            TKit.deleteDirectoryContentsRecursive(outputDir());
-        } else if (ThrowingSupplier.toSupplier(() -> TKit.deleteIfExists(
-                outputBundle())).get()) {
-            TKit.trace(
-                    String.format("Deleted [%s] file before running jpackage",
-                            outputBundle()));
+        if (!noCleanupBeforeExec) {
+            if (isImagePackageType()) {
+                TKit.deleteDirectoryContentsRecursive(outputDir());
+            } else if (ThrowingSupplier.toSupplier(() -> TKit.deleteIfExists(
+                    outputBundle())).get()) {
+                TKit.trace(
+                        String.format("Deleted [%s] file before running jpackage",
+                                outputBundle()));
+            }
         }
 
         Path resourceDir = getArgumentValue("--resource-dir", () -> null, Path::of);
@@ -989,6 +998,7 @@ public final class JPackageCommand extends CommandArguments<JPackageCommand> {
     private boolean suppressOutput;
     private boolean ignoreDefaultRuntime;
     private boolean ignoreDefaultVerbose;
+    private boolean noCleanupBeforeExec;
     private boolean immutable;
     private final Actions prerequisiteActions;
     private final Actions verifyActions;
